rename file for custom error
